+++ test/src/formatter/simple_formatter_test.types.dart
@@ -1,0 +1,3 @@
+part of 'simple_formatter_test.dart';
+
+class _MockAnsiPen extends Mock implements AnsiPen {}